requirements.txt
gradio==3.1.4
transformers
torch
aws-cdk-lib>=2.0.0
constructs>=10.0.0